euci: overload default get and set methods to support types
This drops previous get_t and set_t methods in favor of overloaded get
and set methods.
You can optionally add keyword argument dtype to get method to ensure
that returned data is of specified type. For now only str, int and bool
are supported.
set methods now detects value argument type and behaves according to it.
In general it just converts value to string but in case of bool it
replaces it with defined value.