euci: add option to ensure that tupple is returned
Idea of this is that user can set to configuration file instead of list
option but we want to make sure that even if he does that that we get
tuple of values. This is exactly for that case. Setting list to True
ensures that get never returns anything else then list. Where on the
other hand setting list to False causes it to never return tuple even if
value is lists (in which case first found list is returned).